Rename Concept.ts to ConceptTooltip.tsx
