+++ app/javascript/components/tooltips/index.ts
@@ -1,3 +1,3 @@
 export { MentoredStudent } from './MentoredStudent'
 export { UserTooltip } from './UserTooltip'
-export { ConceptTooltip } from './Concept'
+export { ConceptTooltip } from './ConceptTooltip'
